update_data_dump/add_languages_to_data_dump/detect_languages_in_data_dump/requirements.txt
fasttext==0.9.2
lingua-language-detector==2.0.2
numpy==1.26.4
pybind11==2.12.0
